chore: move preParseDateItems() to CollectionService
